Basics of explore page working again.  no paging or search yet.
A single array of product objects has been replaced with three backbone
collections.  In the dom we now also have three ul elements backed by three
products-list components.  Instead of modifying the products array and
rerendering when switching from e.g. popular to nearby we now hide popular
and show nearby.  This not only means that we can cache the views to avoid
unnecessary re-rendering, we also replaced a single object of mutable state
with three objects which are semantically static.

The backbone collections also provide a much better platform for events.

The framework of the explore page (i.e. the header), now only renders once on
initialize instead of rerendering every time the products change.

The products-list component will now listen to events on its collection instead
of waiting for its owner to manually call render.  The products-list will also
now append views to the dom when models are added to its collection, as opposed
to rerendeing them all every time as it does now.

The search and pagination haven't been updated yet, but they should be working
again soon.